18 November 13:31pm Update
The character now shoots!! I created the bullet in photoshop with a transparent background and imported it into unity, made it a prefab, dragged it into the prefab folder, created a child object for the player to create a fire point. created a Bullet.cs so the bullet spawns on player's direction and object is destroyed after a couple seconds. PlayerController.cs code has the bullet firing upon left click of the mouse.