18 November 14:42pm Update
I now have an enemy that spawns, takes damage when shot, and is destroyed when it is hit 3 times. In the Enemy.cs script I can even get the enemy to flash red on hit. The spawn time and distance isn't exactly right, it also doesn't face the correct direction when it spawns (my fault for using the wrong sprite to create the prefab). But should be easily fixable. Moving the player left to right, jumping, crouching, shooting, platforming, enemy spawning, and a chunk of level to explore, should be enough for the alpha.