feat(animation): scale jump/roll frames
feat(animation): scale jump/roll frames. The scale of the roll animation looked off, added transform property to the jump anim and added keyframes for scaling.
Fix(animation): Fixed the position of the crouch (Hero_Duck_Right) animation in the Sprite Renderer. There was a visible white line behind the character on crouch. It's gone now.