Feat(enemy): EnemyMech shoots bullets 15/12 16:39
Feat(enemy): EnemyMech shoots bullets. I used Photoshop to change the colour of my premade Bullet, and imported the new EnemyBullet into the project. Made it a prefab and gave it a Box Collider and Circle Collider. Created a new EnemyBullet.cs script and assigned it to EnemyBullet. Updated Enemy.cs with new code and new public float. Created an EnemyFirePoint child object to set the bullets to fire from the muzzle. The bullets very in a strange way, it needs some fixing.

    // Enemy shooting
    if (Time.time >= nextEnemyFireTime && player != null)
    {
        nextEnemyFireTime = Time.time + enemyFireRate;

        Vector2 dirToPlayer = (player.position - transform.position).normalized;
        Vector3 spawnPos = enemyFirePoint.position;

        GameObject bullet = Instantiate(enemyBulletPrefab, spawnPos, Quaternion.identity);
        bullet.GetComponent<EnemyBullet>().Initialise(dirToPlayer);
    }
}