Feat(pickups): Added Health PickUps 26/12 11:09
Feat(pickups): Added Health PickUps. Duplicated the Health sprite, brought it into the scene, gave it RigidBody and CircleCollider and made it a Prefab. Gave it a script HealthPickup.cs. Dragged some Health pickups into the scene. Also updated the Enemy.cs script so enemies have a 25% chance to drop Health, this can be adjusted in the Inspector. Fully tested, the Health pick up works and enemies have a chance to drop the Health icon on death.