Fixed(platform): Player moves with platform 16/12 11:52
Fixed(platform): Player moves with platform. I went back and looked at the code in MovingPlatform.cs and realized the 'collision' version of the code was returning errors because I didn't call 'collision' in the method. I re-did the code and it works now. The player when standing on the moving platform will move with the platform. Although, it is not perfect, the player's left and right movement becomes slow and sluggish when on the moving platform. I spent time trying to fix it but couldn't. I've spent enough time on this issue, there are higher priority features to focus on implementing.