Feat(enemy): Added new flying enemy 16/12 16:00
Feat(enemy): Added new flying enemy. FlyingBot added to the project, given colliders, new FlyingBot.cs script, and SpawnManager.cs has it spawn from above the screen. The idea is it is supposed to attack the player in swooping arcs, but it doesn't exactly work the way I had in mind. Still it makes for good variety, keeps the player on their toes.