Feat(scene): Death Zones for falling Player 17/12 15:46
Feat(scene): Death Zones for falling Player. There are gaps in the level with potential for the player to fall. Previously player was falling forever. I created an empty object in the Hierarchy and called it DeathZone. Gave it a Box Collider with IsTrigger = true and altered the size of the object to cover the first gap. I gave it a new DeathZone.cs script. I've playtested it and it works. I dragged it into the Prefab folder because I need at least one more DeathZone.

using UnityEngine;
using UnityEngine.SceneManagement;

public class DeathZone : MonoBehaviour
{
    void OnTriggerEnter2D(Collider2D other)
    {
        if (other.CompareTag("Player"))
        {
            // Restart level
            SceneManager.LoadScene(SceneManager.GetActiveScene().name);
        }
    }
}