Feat(enemy): BossEnemy
Feat(enemy): BossEnemy imported into the project. It has Rigidbody, Box Collider, and new EnemyBoss.cs script. Updated the SpawnManager so EnemyBoss spawns towards the end of the level. It has health, and is destroyed when health is depleted. It's behavior is a little odd though, will require some fine tuning.
BossEnemy.png