Fix(enemy): Boss behaviour consistent 18/12 13:01
Fix(enemy): Boss behaviour consistent.
I created an alternate spawn point for the player in the Inspector to more efficiently test the EnemyBoss encounter and iterated on the size of the boss arena, boss moves on the ground with RigidBody set to Kinematic in the code, boss moves towards player, boss faces player, and shoots at the player consistently. Created a wall at the edge of the boss arena.