Fix(enemy): Enemies shoot through ground 18/12 14:16
Fix(enemy): Enemies shoot through ground.
In order to improve the EnemyBoss encounter I need the player to jump on platforms to avoid getting crushed. But if on a platform and shooting at EnemyBoss, the player can cheese the encounter by just staying there and shooting at will. To fix I deleted the ground check in EnemyBullet.cs and the enemies can now shoot through the ground. Unfortunately, neither for the player or enemies I cannot figure out how to make the bullets visibly pass through the ground tiles. I tried layer sorting and this just seemed to break several things in the game without giving me the desired effect. I've spent long enough on this issue, time to focus on other priorities.