feat(UI): Stage Clear and Restart 18/12 17:25
feat(UI): Stage Clear and Restart.
A couple of moments after EnemyBoss is defeated the words STAGE CLEAR and RESTART appear on the screen. RESTART is clickable and successfully restarts the game. This was created through a Canvas object in the Hierarchy I named StageClearUI, with an image child object and TextMeshPro child object. Button component and new VictoryUI.cs script and updated Update OnBossDefeated() in PlayerMovement triggers the UI and game restart.