feat(music): Music is in the game 19/12 11:08
feat(music): Music is in the game.
A dramatic juxtaposition between the style of music and the style of game.
fix(Hierarchy): I also cleaned up the Hierarchy. It's much tidier now. It was starting to look clogged up.
Fix(level): Made some changes here and there to the level layout and playtested a few times. The game is plenty challenging, but it is possible to beat. I think it's in just about the sweet spot for the old school.