Fix(platform): Free movement for player on MovingPlatform 19/12 17:37
Fix(platform): Free movement for player on MovingPlatform.
After trying different ways to have both free player movement on MovingPlatforms and for the player to move with the platform at the same time, I have run into a wall. I can have one or the other. If the player moves with the platform, the player's movement on the platform is slow and sluggish. If the player has free movement, the player must continually adjust their position on the platform because the platform moves without the player. I chose free movement over sluggishness.
I tried many things to have both movements. I tried adding an OnCollisionStay2D method that didn't work. I tried changing the layers. I tried taking out and putting back in OnCollisionEnter2D and Exit methods. But at least with Rigidbody the player can move freely on the platform. That will have to do.