No fix(Tiles): Transparent for Tiles makes it worse 19/12 17:55
No fix(Tiles): Transparent for Tiles makes it worse 19/12.
Occasionally I can notice a little purple stripe next to the ground Tiles while play testing. This is the purple background of the source png showing. I used photoshop to make the background transparent but as soon as I exported, it created a much more glaring effect of gaps between all the tiles. Changing the size of the tiles also seems ineffective or at least it would be time intensive to figure out exactly what size each tile needs to be to avoid any gaps between them in the game at all.